Refactor Footer component and remove unnecessary code
The Footer component was refactored to remove unnecessary code that was commented out. This improves code readability and maintainability.

Refactor header.css and add new styles

The header.css file was refactored to add new styles for the .list_nav class. This class now displays flex, aligns items and justifies content in the center, and has margin set to 0px. This change enhances the layout and design of the header.

Update Header component and add new logo

The Header component was updated to import a new logo image, controllerlogo.png. This logo is displayed as a link to an external website, tictactoe.anupdhoble.tech. The logo is styled with the .headerlogo class, which sets the width, applies a filter effect, and adds padding. This addition enhances the visual appeal of the header.